Lookup replacement
Changed the way lookup txt files are loaded.
It now loads all .txt files in the lookup folder instead and uses a function call to see wish one to look in.
This will allow for more customization later, and is currently used for the "bitflaglist" parser used for spells and ability lists